removed the last traces of WFocusSet (well, almost, the unused symbol still is in qwindefs.h) and used QWidget::focusWidget() instead.
changed the way clearFocus() works.  now, if you do w->clearFocus()
and the user moves the mouse out of the window and back in, focus may
come back.  or, it may go somewhere else.  but it probably will go
somewhere.

used the above to make QComboBox almost flicker-free and reliably
tabbable.

[imported from CVS by cvs2p4 at 1999/10/31 18:32:52]

[git-p4: depot-paths = "//depot/qt/main/": change = 5460]