improved focus handling when changing tabs; backed out {set,get}Margins
this exposes a bug in either metis or QCheckBox/QButton/QWidget: start
the metis preferences dialog, tab a couple of times so the focus is on
the startup page.  type alt-g or alt-i to switch to the graphics or
grid pages: focus is now on the checkbox on top of the page, but not
visible.  typing tab then shift-tab shows focus, as does moving or
resizing the dialog.

[imported from CVS by cvs2p4 at 1999/10/31 18:44:18]

[git-p4: depot-paths = "//depot/qt/main/": change = 6004]