font, gui style and palette changes should now be handled cleanly and efficiently.
some drawing optimization and a bugfix.  (most mouse and key event
cause double repaint of some items.  bug.  needs to be discussed.)

changed the virtual functions in QListViewItem; createChildren() is
gone, setup() is new and can be used for the same jobs and more;
setSelectable() allows items to be non-selectable but still present;
setExpandable() lets an item be expandable even it it has no children
(in which case the virtaul setOpen() must insert some when called).

new setAllColumnsShowfocus() for the odd list views where all columns
change appearance when the item indicates selectedness or focus.  this
is mostly a drawing optimization.

the dirview example is now broke

[imported from CVS by cvs2p4 at 1999/10/31 18:55:14]

[git-p4: depot-paths = "//depot/qt/main/": change = 6668]