ran qt20fix on qt/src; it made 150-200 changes and as far as I can see, it made all the right changes and no other changes.
renamed QColor::red() to QColor::r(), green() to g() and blue() to b()
because of the conflict with the new colors in QColor.  the colors are
in QColor because my fingers liked them there (QColor::black doesn't
belong in QPaintDevice any more than NoPen does), and because lots of
references to them were in classes that didn't inherit QPD, anyway.
(mostly in owners of widgets/dialogs, but also in static convenience
functions.)

[imported from CVS by cvs2p4 at 1999/10/31 19:52:57]

[git-p4: depot-paths = "//depot/qt/main/": change = 8689]