setEnabled(bool) now deals with child widgets as one expects. This makes it very easy to deal with groupboxes, QScrollView's with childwidgets and all other kind of container widgets. The clumbsy to use and therefore usually ignored (even in Qt itself) function isEnabledToTLW() is deprecated and now equivalent to isEnabled(). Removed some hacks where certain widgets tried to work around the missing hierachical enabling/disabling.
isVisible() now returns whether a widget is visible or not. The
previous semantics ("Would this widget become visible in case its
parent became visible?") was hardly understood by anybody. It was more
or less the direct translation of the internal XWindow mapping state,
which isn't really interesting to the application programmer a

[imported from CVS by cvs2p4 at 1999/10/31 21:58:54]

[git-p4: depot-paths = "//depot/qt/main/": change = 13743]