it used to be that only five of the 15 built-in brushes worked.  it's been like that for years, probably forever. now, all 15 do.
this change also makes the header slightly smaller and faster,
and raises the requirement of vm space in the rip a bit.
shouldn't be a problem even on the oldest printers; drawing a
complex letter requires even more vm.

Qt::CustomPattern still does not work.  will tackle that, too,
but check in this first.

side note: it looks likely that at least one of tests/psoutput
and tests/brushes will produce "unexpected" output on color
postscript printers.  we need to buy one if we want qpsprinter
to work in color.

[git-p4: depot-paths = "//depot/qt/main/": change = 16309]