he new and faster post event system works now (after half a day of debugging, grmblll).
The sendPostedEvent function had and has an invariant: if a QPostEvent struct
is invalid, it *must* not* be referrenced anymore in a local widget list.

This was still the case when invalidating posted ChildInsert events in notify()
when a ChildRemoved event comes along.

Fixed now, please be very careful when optimizing the code any further.

[git-p4: depot-paths = "//depot/qt/main/": change = 21551]