back out almost all of the VMS patch.  kept the new slashify() calls, though, they were clearly good.
cups support in qpd.  not good.  it finds the printers, but nothing
about them.  to get information about printers, we either have to link
in the network extension or use cups' blocking network code.  if we
link in the extensions, we probably don't have to link with libcups at
all.

[git-p4: depot-paths = "//depot/qt/main/": change = 22991]