Redo change 25027. Undo changes 26358 and 26382. Pointers to member functions cannot be cast up and down the inheritance hierarchy that easily. Even the size is not constant! Let's take an example: 	struct Q { 	};
	struct V {
	};

	struct X : public Q, public virtual V {
	};

	typedef void (Q::*QMember)();
	typedef void (X::*XMember)();

Here is what Visual C++ tells us:
	sizeof(QMember) is 4
	sizeof(XMember) is 12

So we'll have to use pointers to pointers to member
functions after all... or we slice them!
Now, we'll have to take care that such pointers to
pointers to members functions are cast properly. This
means using reinterpret_cast<> as far as I can tell...

[git-p4: depot-paths = "//depot/qt/main/": change = 26462]