(Hopefully) last API cleanup iteration for interfaces. - use QGuid instead of QString as interface identifier - make QUnknownInterface pure abstract
Consequences: 
Every interface implementation (aka. component) has to implement queryInterface, addRef and release. This is now exactly what COM does and is probably the only reasonable way for a really open interface architecture (which hides implementation details).
Qt may offer a QUnknownInterface-derived interface which automates queryInterface, addRef and release using an internal list, but this has not very high priority right now. I will rewrite the existing interface implementations (in the Designer) and see if this is really necessary.

[git-p4: depot-paths = "//depot/qt/main/": change = 31723]