Enable incremental parsing for eat_ws() and next_eat_ws().
If speed issues are getting a problem, you could try to avoid
doing the EOF test in eat_ws() and rather enable the parsing
functions to handle the case that eat_ws() was called, but the
character at the cursor position is still a white space character.
(I.e. revert this change and look really careful at the transition
tables and the commands that might be executed twice.)

[git-p4: depot-paths = "//depot/qt/main/": change = 34519]