New paintdevice attribute physical resolution added to all widgets. This is needed to get support for the setResolution() call in QPrinter. We still need to know the real physical resolution of the device in some cases.
Later on (3.1), I'd also like to add the possibility to change the logical resolution 
of any paintdevice, making it a lot easier to achieve wysiwyg with high resolutions.
This would help people doing CAD, DTP or similar things with Qt a lot.

qws and mac: please have a look at the changes I made, especially 
in qprinter_*.cpp

[git-p4: depot-paths = "//depot/qt/main/": change = 36999]