Save C++ source which is edited in the designer in seperate source files and not in the XML file and do the required stuff in uic for that. This makes everything some much easier and cleaner (as the source can be edited also in another editor easily now)
Yes, we are short before a release and I always say: No big changes
(and I keep saying it :-). But this had to be done (and better now than
later) and I discussed it with Matthias. Also the change seems much
bigger than it actually is, because we do exactly the same in Quick since
a long time (designer supported both for a long time now, saving source
in the .ui file and in extra code files). So I basically had to change the language
interface of C++ to say "extra code files please". The rest is exactly the same
as used for a long time in Quick now, so it has been tested a lot and works
very well. 

I also tested with existing C++ designer projects, so backwards compatibility
works too.

[git-p4: depot-paths = "//depot/qt/main/": change = 44817]