Some code cleanup, esp. qlibrary_p.h
Plugins return the version number and threading model of the Qt library used in the ucm_initialize function. If there is any kind of incompatibility cancel the loading. Problem is that the Qt DLL is a different file even for binary compatible versions, so that we get a duplicate of all static objects anyway.

[git-p4: depot-paths = "//depot/qt/main/": change = 45333]