Fix memory corruption. QFontCache::deleteItem was never being called. Instead QCache::deleteItem would be called, which could delete QFontStructs contained in the cache even if their reference count was > 0. Such deletion would result in QFontPrivate::fin dangling pointers to the deleted QFontStructs
On the mac it seems in order for QFontCache::deleteItem to be called when an item is expired from the cache (or the cache is autodeleted) an explicit template specialization is required.

Also on the mac both the QFont and the QFontCache own a reference to any QFontStruct constructed so that QFontCache::deleteItem should call deref when a QFontStruct is removed from the cache.

[git-p4: depot-paths = "//depot/qt/main/": change = 49952]