Fixed a problem with garbage being shown when mode is FixedColumnWidth.
The doc (QTextDocument) width is used to determine how much width 
of the text edit is cleared. The doc width is set to visibleWidth when 
setWrapColumnOrWidth is called. If setWrapColumnOrWidth was called 
before the widget was layouted then the doc width needs to
be updated when the widget is layouted and resized.

[git-p4: depot-paths = "//depot/qt/main/": change = 51907]