This is an attempt to switch from -spec to -platform.
Why are we using -spec in the first place? The configure script
is using -platform and -xplatform (and -xspec does not make much
sense). Actually we ask for a platform (OS + compiler combination)
and get the relevant spec files, so -platform seems to be the right
choice.

Any ideas on this issue? If not, I'll make further changes...

[git-p4: depot-paths = "//depot/qt/main/": change = 54280]