A little cleanup.
Integrate only the file revisions listed in the given change, not up to
head of the file on the branch. Additionally check for out-of-order
integrations by looking to see if the p4 integrate command would
consider a range of revisions, and disallow that.

Autosubmitting is a bad idea, so instead prompt the user to submit, or
to create a changelist he can review and then submit later.

[git-p4: depot-paths = "//depot/qt/main/": change = 56882]