The Q_CHECK_PTR documentation states: 	Q_CHECK_PTR(p) where p is a pointer. 	Writes the warning "In file file.cpp, line 234: 	Out of memory" if p is null. There's no mention of a fatal outcome. So I'm changing from qFatal() to qWarning(). This makes sense anyway. Who needs a program to crash just because there's not enough memory?
Also I don't understand why qt_check_pointer() is a bool
function instead of a void function. There's no need for
a return value, actually Q_CHECK_PTR is a null expression
if QT_CHECK_NULL is not defined so it really should be void.
Mark for change in Qt 4.

Finally I don't understand why the code for Q_CHECK_PTR is:
	(qt_check_pointer((p)==0,__FILE__,__LINE__))
instead of:
	{if ((p)==0) qt_check_pointer(__FILE__,__LINE__);}
which is faster. Mark for change in Qt 4.

[git-p4: depot-paths = "//depot/qt/main/": change = 57014]