Re-did Designer and Assistant interaction.
Assistant now no longer listens on localhost:7358 for new pages. 
(Feature is still there, but not used.) This was a terrible hack
rendering it completly impossible for user to run the Assistant 
from Designer on the same host,	and a potential security hole.

Designer now starts Assistant as a child process, and Assistant
listens on stdin for new pages. It is also possible to send certain
commands to Assistant, like /hide, /show & /goHome...

[git-p4: depot-paths = "//depot/qt/main/": change = 57994]