Aaaargh... Revert previous change.
There's no way we can win against the compiler. It's
using #pragma's to redefine open() to open64(), not
plain #define's.

This means we have to get rid of qplatformdefs.h if
we want to support Large Files. Unless we can somehow
hack the sources not to define _FILE_OFFSET_BITS to 64
except in src/tools/qfile_unix.cpp...

[git-p4: depot-paths = "//depot/qt/main/": change = 61631]