Remove "qindowdefs.h" inclusion. It's a Bad Thing to have tools depend on kernel. And it's also a Bad Thing to include "qindowdefs.h" which includes "qobjectdefs.h" in tools because including it clashes with moc.
It compiles just fine on UNIX this way. I don't know
about Windows and Mac, and I'm sorry if something was
broken on these platforms. If you need functionality
from "qindowdefs.h" in tools, I suggest we move part
of "qindowdefs.h" to some header in tools... Ask me
for qssistance if needed, I'll be all too glad to help
and/or give advice with that.

[git-p4: depot-paths = "//depot/qt/main/": change = 62185]