QSA only: Big change. Finally the designer (MainWindow, etc.) doesn't need to be instantiated anymore to execute a script project. Project, etc. can live alone and only if an error occures and the user wants to debug or the user opens the QSA Developer, the Designer is instantiated as a whole.
This means startup time for QSA enabled apps is much improved (as
the designer is not instantiated on startup) and memory consumtion
is also less until the designer gets instantiated the first time for the
same reasons.

Probably there are some more bugs now. Please let me know if 
you find one.

[git-p4: depot-paths = "//depot/qt/main/": change = 67356]