Clean up the while #ifdef UNICODE/#ifndef Q_OS_TEMP mess.
Use a new macro QT_WA( unicode, ansi ) for the Unicode vs Ansi codeblocks. On CE, we can define it to 'unicode', thus getting completely 
rid of the Ansi block without having to have thousands of ifdefs. On Win32, it's defined to use Unicode or Ansi based on the winVersion.

[git-p4: depot-paths = "//depot/qt/main/": change = 67848]