implemented the Boyle-Moore fast string searching algorithm. This algorithm has a small overhead for building up a skip table, so we use it only in case both strings (the search string and the one to be searched) have a certain length.
Use the new algorithm in replace( QString, QString ), turning it's replacement behaviour 
from quadratic to linear (replacing multiple occurances of a substring in a long string can
be as much as 1000 times faster).

Added a few testcases to test behaviour.

Still need to work on regexp replacing and reverse searching.

[git-p4: depot-paths = "//depot/qt/main/": change = 68192]