re-add the static QMutex to QApplication, and use it as the global Qt library mutex again (instead of the QEventLoop mutex).
QApplication::postEvent is now *completely* threadsafe, obsoleting
QThread::postEvent().  All friends of QApplication::postEvent are
also threadsafe (ie. sendPostedEvents, removePostedEvents, etc.)

done together with lars... and BOY AM I HAPPY :) this has been on my
wishlist for a long time now

make the eventLoop() and setEventLoop() methods in QApplication non-static.
we already have the requirement that QApplication must exist before
creating the eventloop, so this helps re-inforce the idea. (and fix
the reference to QApplication::eventLoop() in the socketnotifier code).

[git-p4: depot-paths = "//depot/qt/main/": change = 68973]