update the docs for QApplication post event functions... they are only threadsafe when building Qt with thread support...
also, the sendPostedEvents() method is threadsafe itself, but the
event delivery mechanism is not.  calling sendPostedEvents() from
a thread that is not the gui thread will cause the events to be
delivered in a thread context that is not the gui thread, which
is very very dangerous.  as such, do not document this function
as threadsafe

[git-p4: depot-paths = "//depot/qt/main/": change = 69071]