use the CDE style if we detect that CDE is running...
since we do not instantiate the styles directly anymore, there is no need
for the #ifndef QT_NO_*STYLE blocks

when reading the font settings from the RESOURCE_MANAGER property, check
to see if the font is an alias, and if it is, use the real font instead.
this works around the fact that CDE creates an 'interface system' alias,
BUT only defines this font in normal weight, and a single size.  Resolving
this alias to the real font allows richtext to work...

[git-p4: depot-paths = "//depot/qt/main/": change = 69160]