Increased the datastream version to 5
The only difference (so far) is the streaming of images (and pixmaps). 

In order to make it possible to read and write null-images (and null
pixmaps ) - something libpng cannot do without error - we now write a
null-image marker (QINT32) before the image data. If the marker is 0,
no image data follows.

The standard conversions image<->pixmap now work free of warnings as
well.

[git-p4: depot-paths = "//depot/qt/main/": change = 69938]