Don't set the d->currPut too early to 0 (in fact, there should be no need to set it to 0 by hand, since it is a guarded pointer). This avoids memory leaks and (what is maybe more important), avoid emitting bogus signals and really stop the operation.
The problem is that the QNetworkProtocol::stop() emits the finished
signal, which in turn then calls the finishedCopy() slot. This
happens before the delete on the d->currPut. So if the d->currPut
is set to 0 in finishedCopy(), it is never deleted and therefore it
continues to operate and work.

This was a problem in the Qtopia desktop (the progress dialog
showed wrong values of an upload was cancelled).

[git-p4: depot-paths = "//depot/qt/main/": change = 71380]