QPrintDialog::getPrinterSetup() doesn't do the right thing.
It creates a QPrintDialog once and forever and then uses it
again and again. Why is it coded this way?

The problem is that QPrintDialog does not have the correct
parent. There is an attempt at faking a parent by calling
adjustPosition() but of course this is not enough to fake
a parent.

The correct way would be to new/delete a QPrintDialog
every time QPrintDialog::getPrinterSetup() is called,
or at least to reparent() a single QPrintDialog.

I'm not fixing this the correct way for now, I don't know
what can happen... Just copying some code from QWidget to
better fake parenting.

[git-p4: depot-paths = "//depot/qt/main/": change = 73657]