support XIM on Solaris (part 2).
the current input method code makes incorrect assumptions about how the user
finishes compositions (ie. by pressing enter).  this is not the case on all
input methods (like Solaris htt), so we need to do this differently.  when
the composition is emptied, we check to see if a pending key event with
keycode == 0 to indicate that a composition has been accepted by the user.
this convention is documented in section 13.21 of the Xlib manual.

we now support XIM on Solaris, which is a common question from many users.

fixes existing tasks: 7036 7261

[git-p4: depot-paths = "//depot/qt/main/": change = 75754]