Revert change 68679.
I wanted to hardwire QMAKE_*_SHLIB so that users cannot
break shared library generation and we don't have plenty
of variables.
Sam suggested not hardwiring it and just adding *_shlib
to CONFIG instead. But then it's just easier for users
to mess up shared library generation using
	CONFIG =
instead of
	CONFIG +=
which is worse.

[git-p4: depot-paths = "//depot/qt/main/": change = 76223]