resolve task 1380.
some changes made in 3.1 after i fixed the task cause QDesktopWidget to
delete all widgets on other screens (something i tried to fixed in change
68035, but apparently broke again).

also, i noticed that the workarea is per screen, so i redid change 73384 to
enable this.

change 73384 also used a QWidget subclass that ignored all insertChild requests
for widget types.  this doesn't work because insertChild is called on
       construction of QObject, at which point isWidgetType returns false.  instead
we remove all children from the QSingleDesktopWidget when it it destructed
(so that the QWidget destructor thinks there are no children).

[git-p4: depot-paths = "//depot/qt/main/": change = 76483]