Simplify #ifdef's.
The Open Group say S_IRWXU is defined in <sys/stat.h> so
always include <sys/stat.h> when Q_NO_SEMAPHORE is defined.

The return value of ftok() and the first argument of semget()
were cast to int which could be an error on some platforms.
Use key_t instead.

Also strerror() is defined in <string.h> so include it.

By the way strerror is not thread-safe. We should be using
thread-safe *_r variants at least in libqt-mt!!!

All this #ifdef'ery should be moved to qplatformdefs.h files
some day...

[git-p4: depot-paths = "//depot/qt/main/": change = 76759]