Back out Trolltech modifications to libpng.
Why were these modifications needed?

If I recall correctly they're supposed to be needed for Metrowerks
CodeWarrior on Mac, and the problem was that CodeWarrior compilers
C files using the C++ compiler. This is not supposed to happen
though. From the Metrowerks manual, "CodeWarrior, Command-Line Tools
Reference":
	A Word about File Extensions 
	Files specified on the command line are identified by
	contents and file extension, as in the CodeWarrior IDE.
	The C/C++ compiler will accept non-standard file
	extensions as source but will emit a warning. By default,
	the compiler assumes that a file with any extensions besides
	.c, .h, .pch is C++ source.
Of course this is revision "030299 cm" from 1999, and it mainly
applies to Windows but something wrong must be going on if .c files
are compiled with the C++ compiler.

Please double-check that there's no error elsewhere and if possible
ask Metrowerks to clarify this issue. Apply modifications again if
there's no other solution.

[git-p4: depot-paths = "//depot/qt/main/": change = 77308]