don't use global objects to initialize the global mutex pools. instead, this should be done in QThread::initialize().
as a result of this, any functions called before QThread::inititalize could
be using a zero qt_global_mutexpool.

so, we should handle this case explicitly.  check if the global mutexpool is
zero, and if so, don't lock.  this is the most efficient for app startup,
which can't be threaded, and then during application runtime, the locking will
occur as expected.

task 6102 first alerted me to this problem.

[git-p4: depot-paths = "//depot/qt/main/": change = 77934]