redo change 72934
the previous fix only handled the DeferredDelete case.  some people are relying
on the behavior that timer events are sent in the previous loop level, which
the previous change broke

i've fixed this properly now by adding a new flag to the eventloop data, which
causes QEventLoop::processEvents to drop out once exit loop is set, but still
allows QEventLoop::processEvents to be called after exitLoop() has been called,
but the event loop hasn't yet returned to the previous loop level.

discussed with and approved by matthias

[git-p4: depot-paths = "//depot/qt/main/": change = 78589]