HAVE_PROTOTYPES does not seem to be needed here. Tested on OpenUNIX with OUDK and GCC.
HAVE_BOOLEAN is needed otherwise:
	jmorecfg.h:227: conflicting types for `typedef int boolean'
	sys/types.h:277: previous declaration as `enum boolean'
Of course this is a very bad hack and should really be fixed
in an upcoming release. Indeed boolean is defined as int when
building the JPEG library and as an enum when building Qt. They're
probably the same size so it shouldn't be much of a problem actually.

One reason for this problem is our use of qplatformdefs.h which
includes <sys/types.h> although it is not needed here.

One problem I have fixing this is that I'm not sure yet how to
detect UnixWare/OpenUNIX reliably. Well of course one may use
__USLC__ to detect the UDK/OUDK compilers and __svr4__ to detect
UnixWare/OpenUNIX when using the GNU compiler. However __svr4__
is probably defined on other System V platforms, isn't it?

[git-p4: depot-paths = "//depot/qt/main/": change = 79044]