p4i integration Integrate 82383 from 3.1 to main:
fix override cursor handling yet again.  some of the logic i reworked
in change 82259 broke cursor setting during widget creation (some
widgets would get the override cursor permanently)  this is because upon
widget creation, setEnabled is called, which tries to set the cursor (to
enforce a certain semantic on X11).  i have fixed this now, and cleaned up
much of the logic to make things much simpler.

this is a continuation of change 82259

[git-p4: depot-paths = "//depot/qt/main/": change = 82384]