manually integrate change 83835:
only lock the mutex once in QComLibrary::createInstanceInternal().

we could run into the problem where one thread creates the
QSettings cache, another thread sees this, the first deletes
the cache, and the other tries to access it (but it is zero,
resulting in a crash).  this is not good, and there is no
point in locking the mutex twice (since we have to do it at
least once to read the cache).

see task 13091

[git-p4: depot-paths = "//depot/qt/main/": change = 83837]