I believe I understand why -lXmu was changed to -lXm - task #11947.
On all UNIX systems libXmu depends on libXt, including HP-UX:
	$ nm -u /usr/contrib/X11R6/lib/libXmu.sl | fgrep Xt
	XtAppSetTypeConverter
	XtAppSetTypeConverter
	XtAppSetWarningMsgHandler
	[...]
However on HP-UX this is not explictly specified by the libray
load list:
	$ ldd -v /usr/contrib/X11R6/lib/libXmu.sl
	$ 
By linking in libXm instead of libXmu, we pull in Motif depencies
which include both libXmu and libXt.

It's better to explictly pull in only libXmu and libXt. We don't
want Qt to link with Motif, especially if Motif is obsoleted and
replaced with GTK.

[git-p4: depot-paths = "//depot/qt/main/": change = 84215]