p4i integration Integrate 84377 from 3.1 to main: Mac specific change.
Make text selection with the mouse more mac like.
This is to handle the case when the user clicks and drags to select text
and the mouse cursor is moved out of the line edit.

On the Mac if the cursor is moved to the right of the click point then
all text to the right of the click point should be selected, and vice
versa for left.

TODO: Make the behavior when the text is too large to be completely visible
in the line edit less windows like and more mac like.

[git-p4: depot-paths = "//depot/qt/main/": change = 84387]