enhancements and additions
added text about adding a custom QDialog to an existing Motif application.

added text about using QMotifDialog to keep an existing Motif dialog
working, but with the Qt API and semantics.  in most cases, these types
of dialogs will probably be removed... but for cases where the dialog
cannot be replaced or rewritten (like if it's a 3rd party add-on),
QMotifDialog needs a way to live with this by providing constructors
with QWidget parents.

another convenience constructor that takes a name/widgetclass/args
would be handy for pre-rolled dialogs.

[git-p4: depot-paths = "//depot/qt/main/": change = 84395]