p4i integration Integrate 84891 from 3.1 to main: Integrate 84890 from 3.0 to 3.1: Cleaner fix for 84705: the problematic case (i.e. the case where you have to disable the read notifier) is only in the ready read signal; if you have a slot connected to the readyRead() signal and you call processEvents() there, you get the bad behaviour that the read notifier fires over and over again and because of the reentrance watcher, nothing is done.
Since this can only happen in slots, but not in general, it is
nicer, more efficient and less error prone to do it just before you
emit the readyRead() signal.

Furthermore, the original 84705 made the qtest application hang on
Linux (for some reason, the QProcess::exited() signal was never
emitted). I have not investigated the reason further, but with this
change it works again.

[git-p4: depot-paths = "//depot/qt/main/": change = 84892]