p4i integration Integrate 85056 from 3.1 to main: On the sco-g++ platform Qt Designer crashes at startup while loading the libwizards.so plugin: 	$ ./bin/designer 	dynamic linker: ./bin/designer: symbol not found: qCleanupImages_wizards__Fv 	Killed 	$
And yet the missing symbol is available in the plugin:
	$ nm plugins/designer/libwizards.so
	...
	[Index]   Value      Size    Type  Bind  Other Shndx   Name
	...
	[799]   |    207120|      90|FUNC |GLOB |0    |14     | qCleanupImages_wizards__Fv
	...
	$ 

The relevant code looks like:
	void qCleanupImages_wizards()
	{
	    ...
	}

	class StaticInitImages_wizards
	{
	public:
	    StaticInitImages_wizards() { qInitImages_wizards(); }
	    ~StaticInitImages_wizards() { qCleanupImages_wizards(); }
	};


It seems that the dynamic linker has problems with the destruction
of static objects. It seems to be looking for functions involved in
destruction of static objects exclusively outside of the plugin,
instead of both outside and within the plugin. This is confirmed by
the fact that using only inline functions for the destruction of the
static object fixes the problem.
Indeed I am able to fix the problem by changing either to:
	class StaticInitImages_wizards
	{
	public:
	    StaticInitImages_wizards() { qInitImages_wizards(); }
	    ~StaticInitImages_wizards() { }
	};
or to:
	inline void qCleanupImages_wizards()
	{
	    ...
	}
The problem reappears if I change to:
	class StaticInitImages_wizards
	{
	public:
	    StaticInitImages_wizards() { qInitImages_wizards(); }
	    ~StaticInitImages_wizards();
	};
	StaticInitImages_wizards::~StaticInitImages_wizards() { }

I don't think there's a technical problem barring the dynamic linker
from looking into the symbols of the plugin itself, because adding
the "static" keyword forces a symbol lookup within the plugin and it
works:
	static void qCleanupImages_wizards()
	{
	    ...
	}

I wasn't able to find a SCO patch to fix the dynamic linker.

[git-p4: depot-paths = "//depot/qt/main/": change = 85057]