Rudementary framework for the MSVC.NET AddIn...
Creates a toolbar with all the standard button from the MSVC6.0 integration. However, I can't seem to remove the label from the buttons, which seems to be a bug in the .NET form classes. Waiting for input on that from Microsoft. Can do some hackish workaround by accessing lowlevel classes, but we've to time to wait for a fix :-)

[git-p4: depot-paths = "//depot/qt/main/": change = 85117]