If the alpha channel is 0, we loose information when calculating the premultiplied alpha channel. So if we have to revert the premultiplication, we have to guess to what value the rgb values should result. It seems to be the safer guess to let them result to (255, 255, 255) than to (0, 0, 0) because QImage::createHeuristicMask() expects this (otherwise all black is removed in the heuristic mask).
This makes the selected icon in a QIconView look better (although
the look is not really optimal, it now looks the same as on Unix,
at least).

At the moment, I don't think it is necessary to store the data of
the original image. Especially when you paint into the pixmaps,
there is no way out of reverting the premultiplied alpha then.

[git-p4: depot-paths = "//depot/qt/main/": change = 85746]