make cursor positioning/movement work together with indic (and other complex) languages.
Did some variable reordering in QTextParagraph to be able to
squeeze the additional pointer and integer I needed in there 
without increasing the size of the object.

The text paragraph has now a pointer to a QTextLayout, that gets used
by the cursors. QTextCursors pointing to the paragraph increase and
decrease a refcount in the paragraph, so we know when we can delete the
text layout object.

[git-p4: depot-paths = "//depot/qt/main/": change = 90667]