rewritten qt_format_text.
It now uses the new QTextLayout classes internally. This finally
fixes the too wide bounding rectangles problems and unifies the
code.

Some paths are a bit slower than in 3.1, but we loose very few against
2.3 and the "richtext case" we had in 3.0/1 is now gone and as fast
as the rest.

With some optimisations in QTextLayout we can hopefully get back
to 3.1 speed for latin text.

[git-p4: depot-paths = "//depot/qt/main/": change = 91767]