p4i integration Integrate 92540 from 3.1 to main: Complete the changes 92520 and 92530 for the general case on Unix: this avoids a SIGPIPE on Unix if a lot of calls to QSocket::writeBlock() are made in sequence.
The change detects in the QSocketDevice::writeBlock() call if the
peer closed the connection. Together with change 92520, this avoids
that QSocket calls ::write() after the RST is already sent.

The method I used is close to what Steven's writes in the table of
page 186 for the scenario that "Our TCP is atively sending data"
and "Peer process crashes": "Peer TCP sends a FIN, which we can
detect immediately using select for readability. If TCP sends
another segment, peer TCP responds with RST. If TCP sends yet
another segment, our TCP sends us SIGPIPE." The last case is what
we want to avoid.

See also task 16211 (the described problem is only for Windows, but
it is also a problem on Unix).

[git-p4: depot-paths = "//depot/qt/main/": change = 92541]