p4i integration Integrate 92861 from 3.1 to main: solve tasks 17060 and 18031:
clipboard operations didn't work between Qt and Motif widgets in the
same application.  this is because qt_xclb_wait_for_event didn't allow
any events to be processed while waiting for data from the clipboard
owner.  this was done for very good reason: processing events could
allow apps to recursively call QClipboard::data() (and, in turn,
qt_xclb_wait_for_event), which would cause lots of bad things to
happen.

however, the likely hood of this happening is very low, with one
exception: the KDE klipper tool.  I HATE THIS DAMN THING.  Lots of
backwards and ugly hacks are necessary to keep this stupid program
working.

if it wasn't for backward compatibility requirements, this change would
have probably been have as big as it is.

note: QEventLoop::processEvents() on X11 now takes an undocumented
flag, 0x08, which stops all timer processing.  This will be introduced
into the public API for QEventLoop in 3.2

discussed/verified with/by matthias

[git-p4: depot-paths = "//depot/qt/main/": change = 92862]