make it compile again on embedded. Mostly works, except for some text which is not drawn.
Required a rather ugly case in QFontEngine::draw() due to
the way the internal graphics driver in QWS is
structured.

The drawText and setFont methods in QGfx are way to high
level. What we would need is a 
drawGlyphs( glyph_t *, positioning info, int num_glyphs )
and a setFontId( QMemoryManager::FontId )

[git-p4: depot-paths = "//depot/qt/main/": change = 93009]