p4i integration Integrate 93018 from 3.1 to main: Integrate 93016 from 3.0 to 3.1: FIXED: QDomNode QDomNode::ownerDocument() works now as the DOM standard specifies.
The QDom classes did QDomNode::ownerDocument() wrong: it always
returned the parent, grandparent, etc. of the node that is of type
document. But this is purely wrong according to DOM specification
(I checked the Apache's Xerxces implementation to make sure that I
understood the specification right). So if you clone a node or if
you import a node to a document, the owner document is now handled
correctly.

Furthermore, the parentNode() implementation was also wrong for the
cases that you created nodes with the QDomDocument::create...()
functions. These functions create the node of the specific type,
but do not insert it into the DOM hierarchy. So the parentNode()
function for these should return a null node. Whereas the
ownerDocument() should return the real owner document.

See also the tests added with change 93011 for concrete examples of
where things went wrong.

See also task 14089.

[git-p4: depot-paths = "//depot/qt/main/": change = 93020]