Introduce new flags for QFileInfo::permissions(): ReadOwner, WriteOwner and ExeOwner -- this has clear semantics (other than the ...User versions). In Qt 4 I plan to change the semantics of the ...User versions to behave like isReadable(), etc. This seems to be the common understanding after discussing this with some people.
But for now this should be sufficient, since one can access now all
interesting data (either via the ...Owner flags or via the
isReadable(), etc. functions). And it doesn't change the behaviour
of any existing code.

The Windows implementation is not complete yet (I have to do this
on Friday together with Andy).

[git-p4: depot-paths = "//depot/qt/main/": change = 95513]