p4i integration Integrate 96118 from 3 to main: Integrate 96117 from 3.1 to 3: This submit contradicts Sams revert 88530, and must be verified by him.
Reason:
When you build Qt multithread shared, and then reconfigure it to multithreaded static, you'll have both
   qt-mt312.lib (dll version)
   qt-mt.lib    (static version) 

Then, when qmake'ing an application with the new static Qt, this routine will detect a library with the version number attached(qt-mt312.lib), and rename the qt link object to match, choosing the dynamic lib instead of the static lib. Only doing the version number checking when QMAKE_QT_DLL is not empty, seems to be a propper fix. Sam?

[git-p4: depot-paths = "//depot/qt/main/": change = 96119]