p4i integration Integrate 99387 from 3 to main: Modify the "rpath" recorded in the Qt library, changing from <build_dir>/lib to <install_dir>/lib for security reasons - task #20098.
Caution!! The effect of this fix is that LD_LIBRARY_PATH
needs to be set to <build_dir>/lib while building Qt. This
was not the case before this fix. As already discussed on
dev@, this shouldn't be such a serious issue because setting
LD_LIBRARY_PATH had always been documented as a requirement.

I'm not modifying Qt 3.1 for now, only Qt 3.2, because that's
nevertheless an important change in the build process. But
considering that's a security issue, I may back port it to
Qt 3.1 if there's not too much negative feedback from Qt 3.2
beta testers.

[git-p4: depot-paths = "//depot/qt/main/": change = 99388]