p4i integration Integrate 100609 from 3 to main: Integrate 100608 from 3.2 to 3: fix a small regression from 3.1.x font matching:
we should prefer exact bitmap font sizes even when the best style is
smoothly scalable.  especially on X11, the exact size will look better
than the scaled version.

this requires us to store the resolution and pixel size for the fonts
that we load, so that we can provide the X server with a more fully
specified XLFD.

see task 22077

[git-p4: depot-paths = "//depot/qt/main/": change = 100610]