remove the qTypeInfo() inline function and introduce the QTypeInfo<T> class.  this allows us to determine type traits at compile time, which gives the complier a better chance to optimize. it also looks tons better.
all of the classes that use the old stuff has been fixed to work with
this new method.  QChar is now declared and complex and movable,
QBitArray is now declared as shared.  others are surely missing and we
will need to find them over time

Some of the nested QRegExpEngine types can't be declared as primitive
types, since the QTypeInfo for them is instantiated before we get a
change to declare our specialization.  The only solution for this is to
move the nested classes/structs out of QRegExpEngine, something that we
should only do if absolutely necessary.	

[git-p4: depot-paths = "//depot/qt/main/": change = 101360]